Changed method of csv parsing to properly return response to frontend
